Demo/requirements.txt
openai>=1.6.0,<2.0.0
python-dotenv>=1.0.0
azure-search-documents>=11.5.0
azure-identity>=1.17.0
pypdf>=4.0.0
requests>=2.31.0
pydantic>=2.0.0
urllib3>=2.0.0
